requirements/base.txt
tensorflow-macos==2.13.0
tensorflow-metal==1.0.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
pillow>=8.3.0
pyyaml>=5.4.0 